+++ src/config/types.ts
@@ -1,0 +1,4 @@
+export interface Configuration {
+  apiAccessToken: string | undefined;
+  baseUrl: string;
+}