+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - Correct supported Node.js version to `>=16`.
  - Send a `User-Agent` header when making Steadybit API calls.
  - Add a `steadybit service open .steadybit.yml` command.
+ - Add the ability to manage configuration profiles via `steadybit config profile`.
 
 ## v0.3.0
  - Rename `contract` to `service definition`